update: add fractional brush sizes and undo/redo counts
Updates brush, path, terrain, and clipboard flows to accept float sizing/strength values (down to 0.5) and adjusts shape/terrain/path iteration logic to use rounded reach while preserving fractional radius behavior. Commands and menus were updated to expose float inputs. Also extends `/we:undo` and `/we:redo` with optional count parameters and fixes history menu rollback behavior by clearing redo entries before mass undo to keep history state consistent.

+++ we_bp/scripts/commands/brush.js
@@ -26,7 +26,7 @@ function brushVariant(name, usePattern) {
             optionalParameters: [
                 { type: CustomCommandParamType.Enum, name: "we:brushshape" },
                 { type: usePattern ? CustomCommandParamType.String : CustomCommandParamType.BlockType, name: "block" },
-                { type: CustomCommandParamType.Integer, name: "radius" },
+                { type: CustomCommandParamType.Float, name: "radius" },
                 { type: CustomCommandParamType.Integer, name: "height" }
             ]
         },
